# perf: remove superfluous call to toLowerCase