@broofa 2023-01-11
> In fact, `toString(16)` seems to already return lower case, but I cannot be sure that this is guaranteed by the spec after reading it

I think this verbiage is the bit you're looking for:

> The digits used in the representation of a number using radix r are taken from the first r code units of "**0123456789abcdefghijklmnopqrstuvwxyz**" in order.

I.e. toString() will always return lowercase strings, so we can remove the `toLowerCase()` call(s) altogether where `bytesToHex` is concerned.