@LinusU 2023-01-11
@broofa hehe, we commented at the same time. Yes, we can skip the call! 🙌