@LinusU 2023-01-11
Name | Before | After
----- | ----- | -----
uuid.stringify() | x 4,012,874 ops/sec ±0.14% | x 3,500,266 ops/sec ±0.17%
uuid.parse() | x 3,773,568 ops/sec ±0.07% | x 3,964,479 ops/sec ±0.13%
uuid.v1() | x 4,003,997 ops/sec ±0.16% | x 6,224,810 ops/sec ±0.18%
uuid.v1() fill existing array | x 9,990,061 ops/sec ±0.16% | x 10,021,774 ops/sec ±0.14%
uuid.v4() | x 22,870,980 ops/sec ±1.75% | x 22,648,164 ops/sec ±1.65%
uuid.v4() fill existing array | x 8,012,358 ops/sec ±0.20% | x 8,100,356 ops/sec ±0.14%
uuid.v3() | x 553,683 ops/sec ±0.27% | x 598,661 ops/sec ±0.27%
uuid.v5() | x 651,688 ops/sec ±0.41% | x 662,520 ops/sec ±0.44%


This change makes the bundle smaller, and the uuid generation faster. On V8 it makes `stringify` slower. I think that this is enough to make this a good change that we want to land.

Potentially we want to add `.toLowerCase()` to `stringify` as a V8-specific optimization, if we feel that it is worth it.