@broofa 2023-01-11
Me calculating how many cpu-decades have been wasted in `toLowerCase()` here...

<image width="400" src="https://media.giphy.com/media/3tEFVAbfzzcwo/giphy-downsized.gif" />