@kkntl 2023-10-11
Apparently the `.toLowerCase()` was introduced in https://github.com/uuidjs/uuid/pull/434 **not** to actually convert to lower case, but as a workaround for a memory issue.

Can you confirm that issue was considered and is not relevant anymore ?

```javascript
// Note: Be careful editing this code!  It's been tuned for performance
// and works in ways you may not expect. See https://github.com/uuidjs/uuid/pull/434
// `toLowerCase` used to fix memory issue caused by concatenation: https://bugs.chromium.org/p/v8/issues/detail?id=3175#c4
```

(I see this statement proven true: https://github.com/uuidjs/uuid/pull/267#issuecomment-378291582:
> ... using tolowercase though might be a little more confusing to anyone else looking at the code.

)