@broofa 2023-10-11
@kkntl Good question!

@LinusU , did you consider the memory consumption issue when putting up this PR?  It looks like that `toLowerCase()` call was actually there to avoid retaining references to the individual substrings that are concatenated together.  (A point I'd completely forgotten about!)

Testing this, it does still appear to be an issue.  See below.

It's worth noting that if this is a regression - as it seems to be - the impact is mitigated to some degree by #600, which short-circuits this project's use of v4() in preference for `crypto.randomUUID()`.   Not that this justifies not fixing this. 🤷 

----

Running this script both with and without the `toLowerCase()` call ...
```javascript
const {v4} = require('uuid');

// Using empty `options` to prevent `uuid` from short-circuiting to
// `crypto.randomUUID()`, which isn't what we want to test
const opts = {}; // empty `options` to

// Call v4() once to prime entropy cache
v4(opts);

// Create memory snapshot here, prior to creating 10K IDs
//
const ids = [];
for (let i = 0; i < 10000; i++) ids.push(v4(opts));

// Create memory snapshot here, after creating 10K IDs, and
// use the "comparison" view in Chrome debugger to inspect
// the difference.

console.log('done');
```

With code on `main` branch (no `toLowerCase()`):
![CleanShot 2023-10-11 at 07 02 35@2x](https://github.com/uuidjs/uuid/assets/164050/b3d208e9-9c9f-489c-a7b7-5c2c79480524)

Adding `toLowerCase()` call back into `stringify()`:
![CleanShot 2023-10-11 at 07 04 24@2x](https://github.com/uuidjs/uuid/assets/164050/0348ee9b-c7f0-49a4-97dc-6f335333013c)